+++ packages/core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @zerodev/sdk
 
+## 5.2.8
+
+### Patch Changes
+
+- Added default only entrypoint hook integration for custom action in kernel v3
+
+## 5.2.7
+
+### Patch Changes
+
+- fix kernelAccountActions typing
+
+## 5.2.6
+
+### Patch Changes
+
+- Fix typing in createKernelAccountClient
+
 ## 5.2.5
 
 ### Patch Changes