+++ plugins/weighted-ecdsa/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @zerodev/weighted-ecdsa-validator
 
+## 5.2.4
+
+### Patch Changes
+
+- Added default only entrypoint hook integration for custom action in kernel v3
+
+## 5.2.3
+
+### Patch Changes
+
+- Update WeigthedECDSAValidator contract address for kernel v3
+
 ## 5.2.2
 
 ### Patch Changes